[#12] View all events ordered as required
Now when the user clicks to view all events, the future events are
ordered in ascending order for the date and names, whereas past events
are ordered in descending order for the date and ascending for their
names. 

Also updated the index.html of the event page to display "Upcoming
events" and "Previous events" separately.