[#19] Redesigned home page for week 6 specification
Now the home page can use Spring Data JPA queries with the addition of
the HomeController.java. 

Moved old home page index.html which was in the
src/main/resources/static/ folder, to the new
src/main/resources/templates/home/ folder so that we can add attributes
to the page.

Updated the a new future event in InitialDataLoader to show that the top
3 query actually only gets the top 3 events (as there are now 4 future
events).

Currently home page displays the upcoming 3 events. Does not yet display
the top 3 venues, but I am working towards that.