[#19] Written tests for the home page controller
Have written tests in HomeControllerTest which test the functions
implemented that populate the home page, with respect to the week 6
requirements.

Also updated the HomePageTest to reflect the way the new layout of the
home page location responds to requests (now HTML instead of JSON).

Also added a function in EventService which could be used in the future
if necessary...